Fix bug with marshling
The `Unmarshal` function used wrong parameter `sDec` instead of the new one `newDec`.  
Second issue was that some tokens are being set with bad base64.  
For example:  
1234 in base64 is: "MTIzNA=="  
If you will remove the "==" in some places you will still get "1234" but in the Go function it remove the "4".
So using "MTIzNA" will result "123" instead of "1234".  
I noticed that some service account tokens have this issue. This is why I added "}" in case it is missing.